requirements/production.txt
-r base.txt
gunicorn==23.0.0
django-storages==1.14.4
boto3==1.35.74